添加定时任务和自动报告生成部分

+++ HeartDataServer/src/main/java/com/qian/Entity/algorithm/DayReport.java
@@ -33,7 +33,7 @@ public class DayReport {
 	private String imgurl;
 	// 1 true 0 false
 	private boolean isUsed;
-	private int healthIndex;
+	private float healthIndex;
 	private int pepoleId;
 	private List<FileReport> fileReportLists;
 }
